backend/requirements.txt
fastapi
uvicorn[standard]
tensorflow
pillow
numpy
google-generativeai
python-dotenv